+++ src/create/viewModel/getDefaultValueForType.js
@@ -1,0 +1,14 @@
+import { Type } from '../../constants/constants.js';
+
+const defaultValues = {
+  [Type.Boolean]: true,
+  [Type.String]: ''
+};
+
+export function getDefaultValueForType(type) {
+  if (defaultValues.hasOwnProperty(type)) {
+    return defaultValues[type];
+  }
+
+  return null;
+}